[web] Move questions logic to its own module
