+++ web/src/components/questions/LuksActivationQuestion.test.jsx
@@ -22,7 +22,7 @@
 import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "@/test-utils";
-import { LuksActivationQuestion } from "@components/storage";
+import { LuksActivationQuestion } from "@components/questions";
 
 let question;
 const answerFn = jest.fn();
