+++ web/src/components/questions/QuestionActions.test.jsx
@@ -22,7 +22,7 @@
 import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "@/test-utils";
-import { QuestionActions } from "@components/core";
+import { QuestionActions } from "@components/questions";
 
 let defaultOption = "handsdown";
