+++ web/src/components/questions/Questions.test.jsx
@@ -25,11 +25,11 @@ import { act, screen, waitFor } from "@testing-library/react";
 import { installerRender } from "@/test-utils";
 import { createClient } from "@client";
 
-import { Questions } from "@components/core";
+import { Questions } from "@components/questions";
 
 jest.mock("@client");
-jest.mock("@components/core/GenericQuestion", () => () => "A Generic question mock");
-jest.mock("@components/storage/LuksActivationQuestion", () => () => "A LUKS activation question mock");
+jest.mock("@components/questions/GenericQuestion", () => () => "A Generic question mock");
+jest.mock("@components/questions/LuksActivationQuestion", () => () => "A LUKS activation question mock");
 
 const handlers = {};
 const genericQuestion = { id: 1, type: 'generic' };
